Further changes as per the review
