--- app/jobs/labor/rate_limit_checker_job.rb
@@ -1,10 +1,0 @@
-module Labor
-  class RateLimitCheckerJob < ApplicationJob
-    queue_as :rate_limit_checker
-
-    def perform(user_id)
-      user = User.find_by(id: user_id)
-      Labor::PingAdmins.call(user) if user
-    end
-  end
-end
